requirements.txt
altair==4.2.0
streamlit==1.19.0
gtts==2.3.1
python-dotenv==0.21.0
requests==2.28.1
matplotlib==3.6.3
numpy==1.24.2
